% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_run.R
\name{import_run}
\alias{import_run}
\title{Import raw data from an AUC run into R}
\usage{
import_run(list_path, dir = NULL)
}
\arguments{
\item{list_path}{The path of the file containing the list of AUC read files.}

\item{dir}{The path to the directory which contains the read files.}
}
\value{
The AUC data from the run in the form of an S4 "Run" object.
}
\description{
This function reads a "list file" which contains a list of AUC scan files.
Each scan file is imported using dcdt::import_scan.
The data is returned in the form of a S4 "Run" object.
}
\examples{
\dontrun{
# Load the run described by the scan files listed in list_file.RA2
auc_run_1 <- dcdt::import_run("path/to/your/list_file.RA2")

# Load the run described by the scan files listed in list_file.RA2 and
# located in C:/Users/haggi/dcdt_data/
auc_run_2 <- dcdt::import_run("path/to/your/list_file.RA2",
  dir="C:/Users/haggi/dcdt_data/")
}

}
