% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_raw}
\alias{plot_raw}
\title{Plot AUC data.}
\usage{
plot_raw(run, start = 1, end = NULL, step = 1, radius_range = NULL)
}
\arguments{
\item{run}{The S4 "Run" object which holds all data from an AUC run.}

\item{start}{The start index to be plotted.}

\item{end}{The end index to be plotted.}

\item{step}{The step size for selecting data between start and end.}

\item{radius_range}{The range of radii to be plotted.}
}
\value{
No return value. This function is used for plotting purposes.
}
\description{
Plot AUC data.
}
\examples{
\dontrun{
# plot every second piece of the raw data from run_1.
plot_raw(run_1, step=2)

# plot all the raw data from run_2, between radius values of 6.25 and 7.1
plot_raw(run_2, radius_range=c(6.25, 7.1))
}

}
