% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readAUC.R
\name{readAUC}
\alias{readAUC}
\title{Import raw AUC data into R}
\usage{
readAUC(file)
}
\arguments{
\item{file}{the location of the file containing the AUC data}
}
\value{
the AUC data in a format usable by R
}
\description{
This function reads data from a X data file (.X) and imports it into R as a X.
}
\examples{
\dontrun{
# Read a X data file
auc.data <- dcdt::readAUC("path/to/your/datafile.X")
}

}
