% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runclass.R
\docType{class}
\name{Run-class}
\alias{Run-class}
\alias{Run}
\title{Definition of an S4 Class for an AUC run.}
\description{
Definition of an S4 Class for an AUC run.
}
\section{Slots}{

\describe{
\item{\code{name}}{character. The name of the run.}

\item{\code{num_scans}}{integer. The number of scans in the run.}

\item{\code{scans}}{list. A list of all the scans in the run (S3 "Scan" objects).}
}}

\examples{
\dontrun{
Create a Run object
run <- new("Run", name = scan$name, num_scans = length(scans), scans = scans)
}

}
